requirements.txt
-i https://pypi.org/simple
blinker==1.6.2 ; python_version >= '3.7'
click==8.1.7 ; python_version >= '3.7'
flask==2.3.3
gunicorn==21.2.0
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
packaging==23.1 ; python_version >= '3.7'
python-dotenv==1.0.0
redis==5.0.0
werkzeug==2.3.7 ; python_version >= '3.8'
